+++ tests/test_WritesRetry.py
@@ -1,4 +1,5 @@
 import unittest
+import time
 
 from urllib3 import HTTPResponse
 from urllib3.exceptions import MaxRetryError
@@ -6,28 +7,58 @@ from urllib3.exceptions import MaxRetryError
 from influxdb_client.client.write.retry import WritesRetry
 
 
+class NonRandomMinWritesRetry(WritesRetry):
+    def _random(self):
+        return 0
+
+
+class NonRandomMaxWritesRetry(WritesRetry):
+    def _random(self):
+        return 1
+
+
 class TestWritesRetry(unittest.TestCase):
     def test_copy(self):
-        retry = WritesRetry(jitter_interval=123, exponential_base=3, max_retry_delay=145)
-        self.assertEqual(retry.jitter_interval, 123)
+        retry = WritesRetry(exponential_base=3, max_retry_delay=145, total=10)
         self.assertEqual(retry.max_retry_delay, 145)
         self.assertEqual(retry.exponential_base, 3)
         self.assertEqual(retry.total, 10)
 
         retry = retry.increment()
-        self.assertEqual(retry.jitter_interval, 123)
         self.assertEqual(retry.max_retry_delay, 145)
         self.assertEqual(retry.exponential_base, 3)
         self.assertEqual(retry.total, 9)
 
         retry = retry.increment()
-        self.assertEqual(retry.jitter_interval, 123)
         self.assertEqual(retry.max_retry_delay, 145)
         self.assertEqual(retry.exponential_base, 3)
         self.assertEqual(retry.total, 8)
 
-    def test_backoff(self):
-        retry = WritesRetry(total=5, backoff_factor=1, max_retry_delay=550)
+    def test_backoff_max_time(self):
+        retry = NonRandomMinWritesRetry(max_retry_time=2)
+        self.assertEqual(retry.is_exhausted(), False)
+        self.assertEqual(retry.get_backoff_time(), 0)
+
+        retry = retry.increment()
+        self.assertEqual(retry.is_exhausted(), False)
+        self.assertEqual(retry.get_backoff_time(), 5)
+
+        retry = retry.increment()
+        self.assertEqual(retry.is_exhausted(), False)
+        self.assertEqual(retry.get_backoff_time(), 10)
+
+        time.sleep(3)
+
+        with self.assertRaises(MaxRetryError) as cm:
+            retry.increment()
+        exception = cm.exception
+        print(exception)
+
+        self.assertEqual("max_retry_time exceeded", exception.reason.args[0])
+
+    def test_backoff_start_range(self):
+        retry = NonRandomMinWritesRetry(total=5, retry_interval=1, exponential_base=2,
+                                        max_retry_delay=550)
         self.assertEqual(retry.total, 5)
         self.assertEqual(retry.is_exhausted(), False)
         self.assertEqual(retry.get_backoff_time(), 0)
@@ -40,22 +71,61 @@ class TestWritesRetry(unittest.TestCase):
         retry = retry.increment()
         self.assertEqual(retry.total, 3)
         self.assertEqual(retry.is_exhausted(), False)
-        self.assertEqual(retry.get_backoff_time(), 5)
+        self.assertEqual(retry.get_backoff_time(), 2)
 
         retry = retry.increment()
         self.assertEqual(retry.total, 2)
         self.assertEqual(retry.is_exhausted(), False)
-        self.assertEqual(retry.get_backoff_time(), 25)
+        self.assertEqual(retry.get_backoff_time(), 4)
 
         retry = retry.increment()
         self.assertEqual(retry.total, 1)
         self.assertEqual(retry.is_exhausted(), False)
-        self.assertEqual(retry.get_backoff_time(), 125)
+        self.assertEqual(retry.get_backoff_time(), 8)
 
         retry = retry.increment()
         self.assertEqual(retry.total, 0)
         self.assertEqual(retry.is_exhausted(), False)
-        self.assertEqual(retry.get_backoff_time(), 550)
+        self.assertEqual(retry.get_backoff_time(), 16)
+
+        with self.assertRaises(MaxRetryError) as cm:
+            retry.increment()
+        exception = cm.exception
+
+        self.assertEqual("too many error responses", exception.reason.args[0])
+
+    def test_backoff_stop_range(self):
+        retry = NonRandomMaxWritesRetry(total=5, retry_interval=5, exponential_base=2,
+                                        max_retry_delay=550)
+
+        self.assertEqual(retry.total, 5)
+        self.assertEqual(retry.is_exhausted(), False)
+        self.assertEqual(retry.get_backoff_time(), 0)
+
+        retry = retry.increment()
+        self.assertEqual(retry.total, 4)
+        self.assertEqual(retry.is_exhausted(), False)
+        self.assertEqual(retry.get_backoff_time(), 10)
+
+        retry = retry.increment()
+        self.assertEqual(retry.total, 3)
+        self.assertEqual(retry.is_exhausted(), False)
+        self.assertEqual(retry.get_backoff_time(), 20)
+
+        retry = retry.increment()
+        self.assertEqual(retry.total, 2)
+        self.assertEqual(retry.is_exhausted(), False)
+        self.assertEqual(retry.get_backoff_time(), 40)
+
+        retry = retry.increment()
+        self.assertEqual(retry.total, 1)
+        self.assertEqual(retry.is_exhausted(), False)
+        self.assertEqual(retry.get_backoff_time(), 80)
+
+        retry = retry.increment()
+        self.assertEqual(retry.total, 0)
+        self.assertEqual(retry.is_exhausted(), False)
+        self.assertEqual(retry.get_backoff_time(), 160)
 
         with self.assertRaises(MaxRetryError) as cm:
             retry.increment()
@@ -64,27 +134,27 @@ class TestWritesRetry(unittest.TestCase):
         self.assertEqual("too many error responses", exception.reason.args[0])
 
     def test_backoff_max(self):
-        retry = WritesRetry(total=5, backoff_factor=1, max_retry_delay=15)\
-            .increment()\
-            .increment()\
-            .increment()\
-            .increment()\
+        retry = WritesRetry(total=5, retry_interval=1, max_retry_delay=15) \
+            .increment() \
+            .increment() \
+            .increment() \
+            .increment() \
             .increment()
 
-        self.assertEqual(retry.get_backoff_time(), 15)
+        self.assertLessEqual(retry.get_backoff_time(), 15)
 
-    def test_backoff_jitter(self):
-        retry = WritesRetry(total=5, backoff_factor=4, jitter_interval=2).increment()
+    def test_backoff_increment(self):
+        retry = WritesRetry(total=5, retry_interval=4).increment()
 
         self.assertEqual(retry.total, 4)
         self.assertEqual(retry.is_exhausted(), False)
 
         backoff_time = retry.get_backoff_time()
         self.assertGreater(backoff_time, 4)
-        self.assertLessEqual(backoff_time, 6)
+        self.assertLessEqual(backoff_time, 8)
 
     def test_backoff_exponential_base(self):
-        retry = WritesRetry(total=5, backoff_factor=2, exponential_base=2)
+        retry = NonRandomMinWritesRetry(total=5, retry_interval=2, exponential_base=2)
 
         retry = retry.increment()
         self.assertEqual(retry.get_backoff_time(), 2)
@@ -145,7 +215,7 @@ class TestWritesRetry(unittest.TestCase):
         response.headers.add('Retry-After', '63')
 
         with self.assertLogs('influxdb_client.client.write.retry', level='WARNING') as cm:
-            WritesRetry(total=5, backoff_factor=1, max_retry_delay=15) \
+            WritesRetry(total=5, retry_interval=1, max_retry_delay=15) \
                 .increment(response=response) \
                 .increment(error=Exception("too many requests")) \
                 .increment(url='http://localhost:9999')